feat!: added parsing for radical indicators ("*")
The component type is extended with an optional field, `is_rad`, which
indicates a component is the radical for this character.

In general the first component that matches the kanji radical is
presumed to be the radical component. However, in some cases
disambiguation is required such as 剪 where ⼑ is the radical, not ⺉.

BREAKING CHANGE: The "*" character is now treated as a special character
meaning "the preceding component is the radical for this character".

BREAKING CHANGE: The `SubComponent` type is now called `BaseComponent`.